+++ test/automigrate/fields_test.clj
@@ -1,7 +1,15 @@
 (ns automigrate.fields-test
-  (:require [clojure.test :refer :all]
-            [clojure.spec.alpha :as s]
-            [automigrate.fields :as fields]))
+  (:require
+    [automigrate.testing-config :as config]
+    [automigrate.testing-util :as test-util]
+    [clojure.test :refer :all]
+    [clojure.spec.alpha :as s]
+    [automigrate.fields :as fields]))
+
+
+(use-fixtures :each
+  (test-util/with-drop-tables config/DATABASE-CONN)
+  (test-util/with-delete-dir config/MIGRATIONS-DIR))
 
 
 (deftest test-validate-fk-options-on-delete
@@ -128,3 +136,154 @@
     (is (false?
           (s/valid? ::fields/validate-fk-options-and-null-on-update {:null false
                                                                      :on-update :set-null})))))
+
+
+(deftest test-fields-alter-column-char-to-int-ok
+  (testing "check generated actions, queries edn and sql from all actions"
+    (is (= {:new-actions (list {:action :alter-column
+                                :changes {:type {:from [:char 256]
+                                                 :to :integer}}
+                                :field-name :num
+                                :model-name :account
+                                :options {:type :integer}})
+            :q-edn [{:create-table [:account]
+                     :with-columns ['(:id :serial)
+                                    '(:num [:char 256])]}
+                    {:alter-table
+                     (list :account
+                       {:alter-column
+                        (list :num :type :integer :using [:raw "num"] [:raw "::"] :integer)})}]
+
+            :q-sql [["CREATE TABLE account (id SERIAL, num CHAR(256))"]
+                    ["ALTER TABLE account ALTER COLUMN num TYPE INTEGER USING num :: INTEGER"]]}
+          (test-util/perform-make-and-migrate!
+            {:jdbc-url config/DATABASE-CONN
+             :existing-actions [{:action :create-table
+                                 :fields {:id {:type :serial}
+                                          :num {:type [:char 256]}}
+                                 :model-name :account}]
+             :existing-models {:account
+                               {:fields [[:id :serial]
+                                         [:num :integer]]}}})))
+
+    (testing "check actual db changes"
+      (testing "test actual db schema after applying the migration"
+        (is (= [{:character_maximum_length nil
+                 :column_default "nextval('account_id_seq'::regclass)"
+                 :column_name "id"
+                 :data_type "integer"
+                 :udt_name "int4"
+                 :is_nullable "NO"
+                 :table_name "account"}
+                {:character_maximum_length nil
+                 :column_default nil
+                 :column_name "num"
+                 :data_type "integer"
+                 :udt_name "int4"
+                 :is_nullable "YES"
+                 :table_name "account"}]
+              (test-util/get-table-schema-from-db
+                config/DATABASE-CONN
+                "account")))))))
+
+
+(deftest test-fields-alter-column-array-with-underscore-ok
+  (testing "check generated actions, queries edn and sql from all actions"
+    (is (= {:new-actions (list {:action :alter-column
+                                :field-name :num-val
+                                :model-name :account
+                                :options {:type :integer}
+                                :changes {:type {:from :char
+                                                 :to :integer}}})
+            :q-edn [{:create-table [:account]
+                     :with-columns ['(:id :serial)
+                                    '(:num-val :char)]}
+                    {:alter-table
+                     (list :account
+                       {:alter-column
+                        (list :num-val :type :integer :using [:raw "num_val"] [:raw "::"] :integer)})}]
+
+            :q-sql [["CREATE TABLE account (id SERIAL, num_val CHAR)"]
+                    ["ALTER TABLE account ALTER COLUMN num_val TYPE INTEGER USING num_val :: INTEGER"]]}
+          (test-util/perform-make-and-migrate!
+            {:jdbc-url config/DATABASE-CONN
+             :existing-actions [{:action :create-table
+                                 :fields {:id {:type :serial}
+                                          :num-val {:type :char}}
+                                 :model-name :account}]
+             :existing-models {:account
+                               {:fields [[:id :serial]
+                                         [:num-val :integer]]}}})))
+
+    (testing "check actual db changes"
+      (testing "test actual db schema after applying the migration"
+        (is (= [{:character_maximum_length nil
+                 :column_default "nextval('account_id_seq'::regclass)"
+                 :column_name "id"
+                 :data_type "integer"
+                 :udt_name "int4"
+                 :is_nullable "NO"
+                 :table_name "account"}
+                {:character_maximum_length nil
+                 :column_default nil
+                 :column_name "num_val"
+                 :data_type "integer"
+                 :udt_name "int4"
+                 :is_nullable "YES"
+                 :table_name "account"}]
+              (test-util/get-table-schema-from-db
+                config/DATABASE-CONN
+                "account")))))))
+
+
+(deftest test-fields-alter-column-array-with-underscore-ok
+  (testing "check generated actions, queries edn and sql from all actions"
+    (is (= {:new-actions (list {:action :alter-column
+                                :field-name :num-val
+                                :model-name :account
+                                :options {:type :char}
+                                :changes {:array {:from "[][]"
+                                                  :to :EMPTY}
+                                          :type {:from :integer
+                                                 :to :char}}})
+            :q-edn [{:create-table [:account]
+                     :with-columns ['(:id :serial)
+                                    '(:num-val :integer [:raw "[][]"])]}
+                    {:alter-table
+                     (list :account
+                       {:alter-column
+                        (list :num-val :type :char
+                          :using [:raw "num_val"] [:raw "::"] :char)})}]
+
+            :q-sql [["CREATE TABLE account (id SERIAL, num_val INTEGER [][])"]
+                    ["ALTER TABLE account ALTER COLUMN num_val TYPE CHAR USING num_val :: CHAR"]]}
+          (test-util/perform-make-and-migrate!
+            {:jdbc-url config/DATABASE-CONN
+             :existing-actions [{:action :create-table
+                                 :fields {:id {:type :serial}
+                                          :num-val {:type :integer
+                                                    :array "[][]"}}
+                                 :model-name :account}]
+             :existing-models {:account
+                               {:fields [[:id :serial]
+                                         [:num-val :char]]}}})))
+
+    (testing "check actual db changes"
+      (testing "test actual db schema after applying the migration"
+        (is (= [{:character_maximum_length nil
+                 :column_default "nextval('account_id_seq'::regclass)"
+                 :column_name "id"
+                 :data_type "integer"
+                 :udt_name "int4"
+                 :is_nullable "NO"
+                 :table_name "account"}
+                {:character_maximum_length 1
+                 :column_default nil
+                 :column_name "num_val"
+                 :data_type "character"
+                 :udt_name "bpchar"
+                 :is_nullable "YES"
+                 :table_name "account"}]
+              (test-util/get-table-schema-from-db
+                config/DATABASE-CONN
+                "account")))))))