+++ CHANGELOG.md
@@ -13,10 +13,12 @@ All notable changes to this project will be documented in this file.
 - Add interval data type.
 - Add parameterised time and timestamp data types with time zone.
 - Add char and varchar data types without parameters.
+- Add ability to create a field with any type as array.
 
 ### Fixed
 
 - Fix validation of parameter for float data type.
+- Fix alter-column action in case column type changing.   
 
 ## 0.1.3 - 2023-10-26